requirements.txt
fastapi[all]
uvicorn
pydantic
python-dotenv
httpx
docker